uploads have a valid flag for each pto-core environment.
- closes #13
- additional configuration parameter `upload_filter`. a dict which is
directly added to the validator's query to check for new uploads.
especially for testing environments this flexibility allows us to
selectively choose which uploads are visible in an environment. in other
words: no need to import all uploads into the action_log if we only want
to analyze certain input formats.